Revert "Move httputil/responseutil to adapters/util and adapters/openrtb_util to exchange/util (#2788)"
This reverts commit 06151091ee480fefa42a25020738f41ec1902333.
