+++ adapters/appnexus/appnexus.go
@@ -13,6 +13,7 @@ import (
 	"github.com/prebid/openrtb/v19/adcom1"
 	"github.com/prebid/openrtb/v19/openrtb2"
 	"github.com/prebid/prebid-server/config"
+	"github.com/prebid/prebid-server/util/httputil"
 	"github.com/prebid/prebid-server/util/randomutil"
 
 	"github.com/prebid/prebid-server/adapters"
@@ -162,11 +163,11 @@ func (a *adapter) MakeRequests(request *openrtb2.BidRequest, reqInfo *adapters.E
 }
 
 func (a *adapter) MakeBids(internalRequest *openrtb2.BidRequest, externalRequest *adapters.RequestData, response *adapters.ResponseData) (*adapters.BidderResponse, []error) {
-	if adapters.IsResponseStatusCodeNoContent(response) {
+	if httputil.IsResponseStatusCodeNoContent(response) {
 		return nil, nil
 	}
 
-	if err := adapters.CheckResponseStatusCodeForErrors(response); err != nil {
+	if err := httputil.CheckResponseStatusCodeForErrors(response); err != nil {
 		return nil, []error{err}
 	}
